Add CI, and point the tooling config at files that exist
No workflow ran in this repository at all, so nothing checked lint,
tests or the bundle on a push. That is how a missing workbox-window
dependency sat on main with a build that could not complete. The new
workflow runs lint, tests and build on pull requests and non-main
branches; the build step matters independently, because tsc passes on a
tree the bundler cannot resolve.

Two config files described a tree that does not exist:

- .prettierrc set tailwindStylesheet to "src/index.css". The stylesheet
  is at src/styles/index.css, so prettier-plugin-tailwindcss could not
  load it. Fixing the path makes the class sorter resolve, which is why
  the next commit is a large formatting-only diff.
- components.json pointed shadcn at "src/index.css" and a
  tailwind.config.js that does not exist under Tailwind v4, and its
  aliases (@/components, @/lib/utils) matched nothing in this tree. Any
  `shadcn add` would have written components into the wrong place.

Also moves isArchived and TAB_FILTER out of HabitsOverview.tsx into
their own module: exporting non-components from a component file breaks
fast refresh, which the new lint step now catches.

Empty leftover directories (src/ui, src/core, src/features/offline and
others) were removed from the working tree. Git does not track empty
directories, so they never existed in the repository -- only on disk.
