+++ .prettierrc
@@ -1,6 +1,8 @@
 {
-  "plugins": ["prettier-plugin-tailwindcss"],
-  "tailwindStylesheet": "src/index.css",
+  "plugins": [
+    "prettier-plugin-tailwindcss"
+  ],
+  "tailwindStylesheet": "src/styles/index.css",
   "printWidth": 80,
   "tabWidth": 2,
   "useTabs": false,
