+++ components.json
@@ -4,17 +4,15 @@
   "rsc": false,
   "tsx": true,
   "tailwind": {
-    "config": "tailwind.config.js",
-    "css": "src/index.css",
+    "css": "src/styles/index.css",
     "baseColor": "zinc",
     "cssVariables": false,
     "prefix": ""
   },
   "aliases": {
-    "components": "@/components",
-    "utils": "@/lib/utils",
-    "ui": "@/components/ui",
-    "lib": "@/lib",
-    "hooks": "@/hooks"
+    "components": "@shared/components",
+    "ui": "@shared/components/ui",
+    "utils": "@shared/utils/utils",
+    "hooks": "@shared/hooks"
   }
 }
